:art: Move server handling to seperate module
